+++ plugin_ide.core/src/com/github/rustdt/ide/core/operations/RustSDKPreferences.java
@@ -2,7 +2,9 @@ package com.github.rustdt.ide.core.operations;
 
 import org.eclipse.core.resources.IProject;
 
+import melnorme.lang.ide.core.LangCore;
 import melnorme.lang.ide.core.operations.ToolchainPreferences;
+import melnorme.lang.ide.core.utils.prefs.IProjectPreference;
 import melnorme.lang.ide.core.utils.prefs.StringPreference;
 import melnorme.lang.tooling.data.LocationValidator;
 import melnorme.lang.tooling.data.ValidationException;
@@ -19,6 +21,11 @@ public interface RustSDKPreferences extends ToolchainPreferences {
 	
 	public static final RustSDKAcessor SDK_PATH_Acessor = new RustSDKAcessor();
 	
-	public static final StringPreference SDK_SRC_PATH = new StringPreference("sdk_src_path", "");
+	public static final IProjectPreference<String> SDK_SRC_PATH2 = new StringPreference(LangCore.PLUGIN_ID, 
+		"sdk_src_path", "", ToolchainPreferences.USE_PROJECT_SETTINGS)
+			.getProjectPreference();
+	
+	public static final IProjectPreference<String> RACER_PATH = new StringPreference(LangCore.PLUGIN_ID, 
+		"racer_path", "", ToolchainPreferences.USE_PROJECT_SETTINGS).getProjectPreference();
 	
 }